+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   mode)
 - [#441] Fixed an issue where the preview object pickable status could get out of sync with the original
 - [#444] Fixed an issue where the preview system broke drag-and-drop of materials onto the scene view
+- [#444] Fixed an issue where the preview system broke drag-to-select in the scene view
 
 ### Changed
 - [#451] Changed preview system to apply to more cameras. Specifically, we now handle all cameras with no render